requirements.txt
authlib==1.0.1
confluent_kafka==1.8.2
fastapi==0.75.1
fastapi_mail==1.0.8
jose==1.0.0
nest_asyncio==1.5.5
numpy==1.22.1
passlib==1.7.2
Pillow==9.1.0
pydantic==1.9.0
pydantic_choices==0.1.1
PyJWT==2.3.0
pytest==7.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.27
starlette==0.17.1
uvicorn==0.17.6
